+++ src/doomstat.h
@@ -539,6 +539,7 @@ extern int a11y_invul_colormap;
 
 extern int no_hor_autoaim;
 extern int switch_on_pickup;
+extern int skip_ammoless_weapons;
 extern int always_bob; // CFG-Only
 
 enum {