+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## New Features
 
 - **_Freecam_**
+- **_Prev/Next [Weapon Buttons] Skip Ammoless Weapons_** setting
 - **_'BABYMODE'_ cheat**, to toggle ITYTD benefits
 - **Setting to play `DSNOWAY` instead of `DSOOF` when failing to use key-locked triggers**